Add acronym find function and module file
A new function named 'findAcronym' has been introduced in the 'acronymOfWorks.go' file. This function takes two parameters - a slice of words and the first letter of an acronym, and checks if the acronym can be formed from the first letters of the words in the provided slice. Also, added 'go.mod' file to define the module's path